LayoutTests/imported/w3c/web-platform-tests/streams/piping/then-interception.any-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: Attempted to assign to readonly property.

Harness Error (FAIL), message = Unhandled rejection: Attempted to assign to readonly property.

TIMEOUT piping should not be observable Test timed out
NOTRUN tee should not be observable 
